add my list button component
